simplify and consolidate command-{result, results, types}.ts
