+++ apps/cli/src/core/commands/init.ts
@@ -34,7 +34,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { fileURLToPath } from 'url';
 import { colors } from '../io/cli-colors.js';
-import { CommandResults } from '../command-results.js';
+import { CommandResults } from '../command-types.js';
 import { CommandBuilder } from '../command-definition.js';
 import { BaseOptionsSchema, withBaseArgs } from '../base-options-schema.js';
